backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioredis==2.0.1
aiosignal==1.4.0
aiosqlite==0.19.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.7.8
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
base58==2.1.1
bcrypt==4.1.2
billiard==4.2.1
bitarray==3.7.0
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ckzg==1.0.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
cytoolz==1.0.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.104.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
hexbytes==0.3.1
hf-xet==1.1.8
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.34.4
hyperon==0.2.2
idna==3.10
ipfshttpclient==0.8.0a2
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jwt==1.4.0
kombu==5.5.4
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.6.4
netaddr==1.3.0
packaging==25.0
parsimonious==0.10.0
passlib==1.7.4
Pillow==10.1.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
py-ecc==8.0.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyunormalize==16.0.0
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rlp==4.1.0
rpds-py==0.27.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tokenizers==0.21.4
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
varint==1.0.2
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
web3==6.11.3
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
duckduckgo-search
deepface
opencv-python
tensorflow